www-plugins/web-eid: Bump to 2.4.0
Update to web-eid-2.4.0.

Also, instead of using git fetch use the more common SRC_URI
approach including the pinned library versions this allows
ensuring the authenticity of the sources for non -9999 builds.

Pull also a pinned version of the XPI for firefox
from mozilla's website.

Finally fix the plugin installation for Chromium and the native
messaging app installation for both Chrome and Chromium.

Chromium expects the json file for the plugin to be present at
/usr/share/chromium/extensions/ instead of
/usr/share/google-chrome/extensions/

Also to be able to interact with the native app, the eu.webeid.json
must be installed on /etc/opt/chrome/native-messaging-hosts/ (chrome)
and on /etc/chromium/native-messaging-hosts/ (chromium).

I do not have chrome to test but the plugin seems to work correctly on
Firefox (after enabling it) and on Chromium (where it is installed
automatically).

Fixes #60
Closes #62

Signed-off-by: Francisco Blas Izquierdo Riera (klondike) <klondike@gentoo.org>